+++ AP_Assignment_3/Row.java
@@ -1,15 +1,15 @@
 import java.util.ArrayList;
 
 public class Row {
-    private static ArrayList<float[][]> R = new ArrayList<float[][]>();
+    private static ArrayList<double[][]> R = new ArrayList<double[][]>();
     
-    Row(float[][] r,int rows, int columns){
+    Row(double[][] r,int rows, int columns){
         if(rows == 1){
             R.add(r);
         }
     }
 
-    public static boolean getRow(float[][] m){
+    public static boolean getRow(double[][] m){
         return R.contains(m);
     }
 }
